Sat, Oct 14, 2023, 4:20 AM +05:00

+++ readability/readability.c
@@ -28,12 +28,12 @@ int main(void)
     // Coleman-Liau index formula
 
     //
-    int L = (letter_count / word_count) * 100;
-    printf("L: %i\n", L);
+    float L = (letter_count / word_count) * 100;
+    printf("L: %f\n", L);
 
     //
-    int S = (sentence_count / word_count) * 100;
-    printf("S: %i\n", S);
+    float S = (sentence_count / word_count) * 100;
+    printf("S: %f\n", S);
 
     int index = 0.0588 * L - 0.296 * S - 15.8;
     printf("Grade %i\n", index);
